CEDS-2964 Add test for Timeout Dialog component
